reg gpu kernel
